requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==3.0.1
nltk==3.8.1
